@@ -6,7 +6,7 @@ version: 1.0.0+1
 publish_to: 'none'
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
   flutter: ">=1.12.0"
 
 dependencies:
@@ -23,7 +23,7 @@ dev_dependencies:
 
   pdftron_flutter:
     path: ../
-  permission_handler: '3.0.1'
+  permission_handler: '8.0.0+2'
 
 # For information on the generic Dart part of this file, see the
 # following page: https://www.dartlang.org/tools/pub/pubspec
