@@ -1,13 +1,13 @@
 name: pdftron_flutter
 description: A convenience wrapper for building Flutter apps with PDFTron mobile SDK.
-version: 1.0.0-beta.1
+version: 1.0.0-beta.2
 homepage: https://www.pdftron.com
 repository: https://github.com/PDFTron/pdftron-flutter
 issue_tracker: https://github.com/PDFTron/pdftron-flutter/issues
 documentation: https://github.com/PDFTron/pdftron-flutter/blob/master/API.md
 
 environment:
-  sdk: ">=2.0.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
   flutter: ">=1.12.0"
 
 dependencies:
